Create a basic popup for displaying IP data on history pages
Add a button next to each anonymous user link. On clicking the button
a popup appears, which queries the API and displays the mock data.
The query is only performed once per popup.

The feature is only available on history pages, for users with the
'ipinfo' right.

Timeouts, errors and loading bars are not yet displayed.

Bug: T260604
Change-Id: I039837480835f00af57418a32d92aa8d78f0c4b9