Allow LogHandler to return IP information for log target
After this the log endpoint can return information for the performer,
the target, or both, given a log ID. It will return data for
whichever are valid IP addresses. If neither is a valid IP address,
it will return an error.

The IP addresses are now returned in the info object, so that the
client can work out which information object to display.

Bug: T263635
Change-Id: I5ed5523114b6086841354e37c3699959b639819f