Control access to log target
In T263635, the target of a log entry was added to the IP Info response
but access was not checked to determine if the user can see the IP
address of the target.

This change adds that check as well as a check for the log type. Users
should now only be able to get the ipinfo for the performer or target
they are able to access.

Bug: T264879
Change-Id: I87ae446061f20cc74d7146e91a26f318dc2edc50