services: Fix IPInfoGeoIP2InfoRetriever
In I448e98f41cc, we changed the IPInfoGeoIP2InfoRetriever factory
function to return an instance of GeoIp2EnterpriseInfoRetriever if the
$wgIPInfoGeoIP2EnterprisePath config variable exists. However, the
config always exists as it is set to false in extension.json and
therefore the factory function will only ever return an instance of
GeoIp2EnterpriseInfoRetriever.

Change the IPInfoGeoIP2InfoRetriever factory function to check if the
config variable is truthy.

Bug: T288933
Change-Id: I6462684bfc54662ce896cc7dcbf5344f54e3a2fb