logging: Combine implementation and interface
The original intent behind introducing the Logger interface was that the
destination of the log entries might change in future. However, the
primary goal of the interface should be to capture the boundary of the
domain, which it did not: regardless of the destination of the log
entries, we should debounce only the view_accordion and view_popup
actions.

One way of capturing the above would be to create:

1. A LoggerDestination interface, implementations of which can be used
   to create a log entry and to check whether a specific log entry
   should be debounced

2. Logger, which collaborates with an implementation of
   LoggerDestination

The above is desirable as it minimises the number of responsibilities of
the Logger and LoggerDestination. However, this comes at the cost of
increased complexity. As of writing, that cost is not worth paying.

Changes:

* Combine the Logger interface and DebouncingLogger implementation

* Add documentation

* Introduce Logger::DEFAULT_DEBOUNCE_DELAY, which is the default value
  for the debounce delay

* Rename and update DebouncingLoggerTest

Additional changes:

* Assert that the delay is positive

Bug: T294680
Change-Id: I051d0c1e708a1dc3de8f9e9980cd34b654e99c03