Disallow API access when feature is not beta enabled
If a user agreed to the terms of use and then disabled the beta
feature, they could still navigate to the API url and view ip info.
Disable that access if BetaFeatures is installed but the beta
feature isn't enabled, no matter if the user has agreed to the terms
of use.

Bug: T299050
Change-Id: I2eec0c14754fd020386267f71702c8e7fd9838d2