Pass a full UserIdentity to InfoRetrievers
In order to allow displaying IP info for temporary users, InfoRetrievers
will need to be able to operate on both temporary and anonymous users.
So, pass a full UserIdentity to them, instead of just an IP address.

This patch does not actually add temporary user handling to
InfoRetrievers - that will come as a follow-up.

Bug: T349715
Change-Id: I70ec861067670e0e5f3a9f90e2ae7ab66d768c96