Make ContributionInfoRetriever work for temporary users
ContributionInfoRetriever currently only supports anonymous users, since
it treats the user name as an IP address and uses it to fetch edit data.
So, have it lookup the actor ID for the input user and use the actor ID
in lookups, which works both for anonymous and temporary users.

Convert the existing unit test for this class into an integration test,
so that the DB queries issued by the class can be tested.

Bug: T349715
Change-Id: I8c9a5f408efa7cf2111f265178458bce8e90ed2c