Introduce TempUserIPLookup service
Now that InfoRetrievers accept an UserIdentity that may either be an
anonymous or a temporary user, they need a mechanism to obtain IP
address data for a temporary user. Introduce a service class that
encapsulates fetching this data from CheckUser. This service class will
be integrated with InfoRetrievers that need it in a follow-up patchset.

Bug: T349715
Change-Id: I85306c56d05c5fcc419805ec934e7fae545700a1