Introduce InfoRetriever to fetch unique IP address count for temp users
Per the task requirements, we will need to display the count of unique
IP addresses associated with a given temporary user in their IPInfo
infobox. Do this by adding a new InfoRetriever that fetches the
requisite data from the newly added TempUserIPLookup service.

This does not yet change the UI in any way, since we do not yet pass
temporary users to InfoRetrievers, nor have we made the frontend aware
of the data that may have been returned by this InfoRetriever. That will
all be done in follow-up patchsets.

Bug: T349715
Change-Id: I1da3d1b8541e3d7e69f88dd9254daf57e5b87551