Add integration test for RevisionHandler
Although RevisionHandlerTest is nominally an integration test, it mocks
most of its dependencies. So, convert it into an unit test instead and
add a more high-level integration test so that we can have more confidence
when adding temporary user support to this handler.

Move the MaxMind test DBs used by browser tests to tests/fixtures so that
it's clearer that they are also used by PHPUnit integration tests.

Bug: T349715
Change-Id: I97fdff81eaf263524be1b1a77895f77ae97eb60b