Add CheckUser/AbuseFilter-lookup endpoint for temp accounts with no revisions
Why:
A user can commit a disallowed action that results in an account being
made but no revisions being associated with the account. Since IPInfo
uses the revision id to get an IP to look up, it cannot look up IP information
for these accounts which may have instead logged the action in CheckUser
or AbuseFilter. These accounts and the IP information associated with them
are still of interest to functionaries.

In these cases, IPInfo should look up the user in those tables to try
and pull an IP to look up.

What:
- Add a new endpoint that takes the username of the account with no revisions
  and uses it to look up the actor in CheckUser events & logs and AbuseFilter logs
  if the extensions are installed

Bug: T380221
Change-Id: I278adc9984353c6a8f5870e4d0e985e367fd58ca