Remove IPInfoHookRunner service
The HookRunner class is not a service, it has no state and does not
provide additional functionality.
It is a lightweight class, that is not expensive to create, there is no
need to cache it in the request in the service container.
The HookRunner class exists to avoid wrong written function name and
also to provide and check type declaration to match the hook interfaces.
Inject the HookContainer instead and create in the class that runs hooks

Change-Id: Ia5bbfd3536282f1c968d15871b9a20037224fbeb